refactored routing factory
