+++ ellar/core/routing/__init__.py
@@ -1,5 +1,10 @@
 from .app import ApplicationRouter
+from .builder import RouterBuilder, get_controller_builder_factory
+from .factory import ControllerRouterFactory
 
 __all__ = [
     "ApplicationRouter",
+    "ControllerRouterFactory",
+    "get_controller_builder_factory",
+    "RouterBuilder",
 ]
