+++ tests/test_controller/test_controller_decorator.py
@@ -5,7 +5,7 @@ from ellar.common.constants import (
     CONTROLLER_CLASS_KEY,
     CONTROLLER_OPERATION_HANDLER_KEY,
 )
-from ellar.common.routing import ControllerRouterFactory
+from ellar.core.routing import ControllerRouterFactory
 from ellar.di import has_binding, is_decorated_with_injectable
 from ellar.reflect import reflect
